NOJIRA: Addressing code review comments
* Schema is now a property of the settings' models.
* Refactored app.js - split into multiple entities.
* Refactored pcp/js/index.js - split into multiple entities.
* Decouple gpiiConnector and psp components.
* Updated license information.
* Renamed IPC events.
* Improved documentation.
