+++ package.json
@@ -28,6 +28,6 @@
     "infusion": "3.0.0-dev.20170830T182157Z.392b2f8",
     "node-jqunit": "1.1.7",
     "request": "2.69.0",
-    "winstrap": "^0.5.11"
+    "winstrap": "0.5.11"
   }
 }
