+++ src/app.js
@@ -12,21 +12,20 @@ https://github.com/GPII/universal/blob/master/LICENSE.txt
 */
 "use strict";
 
-var electron = require("electron");
-var fluid = require("infusion");
-var gpii = fluid.registerNamespace("gpii");
-var path = require("path");
+var fluid   = require("infusion");
+var gpii    = fluid.registerNamespace("gpii");
+var path    = require("path");
 var request = require("request");
-var os = require("os");
 
-var BrowserWindow = electron.BrowserWindow,
-    Menu = electron.Menu,
-    Tray = electron.Tray,
-    globalShortcut = electron.globalShortcut,
-    ipcMain = electron.ipcMain,
-    systemPreferences = electron.systemPreferences;
-var ws = require("ws");
+
+require("./gpiiConnector.js");
+require("./menu.js"); // menuInApp, menuInAppDev
+require("./tray.js");
+require("./pcp.js");
+require("./waitDialog.js");
+
 require("./networkCheck.js");
+
 
 /**
  * Promise that resolves when the electron application is ready.
@@ -41,89 +40,6 @@ gpii.app.electronAppListener = function () {
     gpii.app.appReady.resolve(true);
 };
 require("electron").app.on("ready", gpii.app.electronAppListener);
-
-
-/**
- * Responsible for creation and housekeeping of the connection to the PCP Channel WebSocket
- */
-fluid.defaults("gpii.app.gpiiConnector", {
-    gradeNames: "fluid.component",
-
-    // Configuration regarding the socket connection
-    config: {
-        gpiiWSUrl: "ws://localhost:8081/pcpChannel"
-    },
-
-    members: {
-        socket: "@expand:gpii.app.createGPIIConnection({that}.options.config)"
-    },
-
-    events: {
-        onPreferencesUpdated: null
-    },
-
-    listeners: {
-        "onDestroy.closeConnection": {
-            listener: "{that}.closeConnection"
-        }
-    },
-
-    invokers: {
-        registerPCPListener: {
-            funcName: "gpii.app.registerPCPListener",
-            args: ["{that}.socket", "{that}", "{arguments}.0"]
-        },
-        updateSetting: {
-            funcName: "gpii.app.gpiiConnector.updateSetting",
-            args: ["{that}.socket", "{arguments}.0"]
-        },
-        updateActivePrefSet: {
-            funcName: "gpii.app.gpiiConnector.updateActivePrefSet",
-            args: ["{that}.socket", "{arguments}.0"]
-        },
-        closeConnection: {
-            this: "{that}.socket",
-            method: "close",
-            // for ref https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent#Status_codes
-            args: [1000]
-        }
-    }
-});
-
-/**
- * Sends setting update request to GPII over the socket.
- *
- * @param socket {Object} The already connected WebSocket instance
- * @param setting {Object} The setting to be changed
- * @param setting.path {String} The id of the setting
- * @param setting.value {String} The new value of the setting
- */
-gpii.app.gpiiConnector.updateSetting = function (socket, setting) {
-    var payload = JSON.stringify({
-        path: ["settingControls", setting.path],
-        type: "ADD",
-        value: setting.value
-    });
-
-    socket.send(payload);
-};
-
-
-/**
- * Send active set change request to GPII.
- *
- * @param socket {ws} The already connected `ws`(`WebSocket`) instance
- * @param newPrefSet {String} The id of the new preference set
- */
-gpii.app.gpiiConnector.updateActivePrefSet = function (socket, newPrefSet) {
-    var payload = JSON.stringify({
-        path: ["activeContextName"],
-        type: "ADD",
-        value: newPrefSet
-    });
-
-    socket.send(payload);
-};
 
 /*
  ** Component to manage the app.
@@ -140,9 +56,19 @@ fluid.defaults("gpii.app", {
     },
     // prerequisites
     components: {
+        pcp: {
+            type: "gpii.app.pcp",
+            createOnEvent: "onPrerequisitesReady",
+            options: {
+                model: {
+                    keyedInUserToken: "{app}.model.keyedInUserToken"
+                }
+            }
+        },
         gpiiConnector: {
             type: "gpii.app.gpiiConnector",
             createOnEvent: "onPrerequisitesReady",
+            priority: "after:pcp",
             options: {
                 listeners: {
                     "onPreferencesUpdated.updateSets": {
@@ -152,21 +78,35 @@ fluid.defaults("gpii.app", {
                 }
             }
         },
-        pcp: {
-            type: "gpii.app.pcp",
+        /*
+         * A helper component used as mediator for handling communication
+         * between the PCP and gpiiConnector components.
+         */
+        pcpMediator: {
+            type: "fluid.component",
             createOnEvent: "onPrerequisitesReady",
+            priority: "after:gpiiConnector",
             options: {
-                model: {
-                    keyedInUserToken: "{app}.model.keyedInUserToken"
-                },
                 listeners: {
-                    "onCreate.registerWithConnector": {
-                        funcName: "{gpiiConnector}.registerPCPListener",
-                        args: "{that}"
+                    "{pcp}.events.onSettingAltered": {
+                        listener: "{gpiiConnector}.updateSetting",
+                        args: ["{arguments}.0"]
+                    },
+                    "{pcp}.events.onActivePreferenceSetAltered": {
+                        listener: "{gpiiConnector}.updateActivePrefSet",
+                        args: ["{arguments}.0"]
+                    },
+
+                    "{gpiiConnector}.events.onPreferencesUpdated": {
+                        listener: "{pcp}.notifyPCPWindow",
+                        args: ["onPreferencesUpdated", "{arguments}.0"]
+                    },
+                    "{gpiiConnector}.events.onSettingUpdated": {
+                        listener: "{pcp}.notifyPCPWindow",
+                        args: ["onSettingUpdated", "{arguments}.0"]
                     }
                 }
-            },
-            priority: "after:gpiiConnector"
+            }
         },
         tray: {
             type: "gpii.app.tray",
@@ -281,18 +221,6 @@ gpii.app.fireAppReady = function (fireFn) {
 };
 
 /**
-  * Refreshes the task tray menu for the GPII Application using the menu in the model
-  * @param tray {Object} An Electron "Tray" object.
-  * @param menuTemplate {Array} A nested array that is the menu template for the GPII Application.
-  * @param events {Object} An object containing the events that may be fired by items in the menu.
-  */
-gpii.app.updateMenu = function (tray, menuTemplate, events) {
-    menuTemplate = gpii.app.menu.expandMenuTemplate(menuTemplate, events);
-
-    tray.setContextMenu(Menu.buildFromTemplate(menuTemplate));
-};
-
-/**
   * Keys into the GPII.
   * Currently uses an url to key in although this should be changed to use Electron IPC.
   * @param token {String} The token to key in with.
@@ -352,299 +280,6 @@ gpii.app.exit = function (that) {
     } else {
         gpii.app.performQuit();
     }
-};
-
-fluid.registerNamespace("gpii.app.pcp");
-
-/**
- * Sends a message to the given window
- *
- * @param pcpWindow {Object} An Electron `BrowserWindow` object
- * @param messageChannel {String} The channel to which the message to be sent
- * @param message {String}
- */
-gpii.app.pcp.notifyPCPWindow = function (pcpWindow, messageChannel, message) {
-    if (pcpWindow) {
-        pcpWindow.webContents.send(messageChannel, message);
-    }
-};
-
-/**
-* Get the position of `Electron` `BrowserWindows`
-* @param {Number} width The current width of the window
-* @param {Number} height The current height of the window
-* @return {{x: Number, y: Number}}
-*/
-gpii.app.getWindowPosition = function (width, height) {
-    var screenSize = electron.screen.getPrimaryDisplay().workAreaSize;
-    return {
-        x: screenSize.width - width,
-        y: screenSize.height - height
-    };
-};
-
-
-/**
- * Creates an Electron `BrowserWindow` that is to be used as the PCP window
- *
- * @param {Object} windowOptions Raw options to be passed to the `BrowserWindow`
- * @returns {Object} The created Electron `BrowserWindow`
- */
-gpii.app.pcp.makePCPWindow = function (windowOptions) {
-    // TODO Make window size relative to the screen size
-    var pcpWindow = new BrowserWindow(windowOptions);
-
-    var url = fluid.stringTemplate("file://%gpii-app/src/pcp/index.html", fluid.module.terms());
-    pcpWindow.loadURL(url);
-
-    gpii.app.pcp.hidePCPWindow(pcpWindow);
-
-    return pcpWindow;
-};
-
-/**
- * A function which should be called to init various listeners related to
- * the PCP window.
- * @param pcp {Object} The `gpii.app.pcp` instance
- */
-gpii.app.initPCPWindowListeners = function (pcp) {
-    var pcpWindow = pcp.pcpWindow;
-    pcpWindow.on("blur", function () {
-        pcp.hide();
-    });
-
-    electron.screen.on("display-metrics-changed", function (event, display, changedMetrics) {
-        if (changedMetrics.indexOf("workArea") > -1) {
-            var windowSize = pcpWindow.getSize(),
-                contentHeight = windowSize[1];
-            pcp.resize(contentHeight, true);
-        }
-    });
-};
-
-/**
- * This function checks whether the PCP window is shown.
- * @param pcpWindow {Object} An Electron `BrowserWindow`
- * @return {Boolean} `true` if the PCP window is shown and `false` otherwise.
- */
-gpii.app.pcp.isPCPWindowShown = function (pcpWindow) {
-    var position = pcpWindow.getPosition(),
-        x = position[0],
-        y = position[1];
-    return x >= 0 && y >= 0;
-};
-
-/**
- * Shows the PCP window in the lower part of the primary display and focuses it.
- * Actually, the PCP window is always shown but it may be positioned off the screen.
- * This is a workaround for the flickering issue observed when the content displayed in
- * the PCP window changes. (Electron does not rerender web pages when the
- * `BrowserWindow` is hidden).
- * @param pcpWindow {Object} An Electron `BrowserWindow`.
- */
-gpii.app.pcp.showPCPWindow = function (pcpWindow) {
-    var screenSize = electron.screen.getPrimaryDisplay().workAreaSize,
-        windowSize = pcpWindow.getSize(),
-        windowX = screenSize.width - windowSize[0],
-        windowY = screenSize.height - windowSize[1];
-    pcpWindow.setPosition(windowX, windowY);
-    pcpWindow.focus();
-};
-
-/**
- * Hides the PCP window by moving it to a non-visible part of the screen. This function
- * in conjunction with `gpii.app.pcp.showPCPWindow` help avoid the flickering issue when
- * the content of the PCP window changes.
- * @param pcpWindow {Object} An Electron `BrowserWindow`.
- */
-gpii.app.pcp.hidePCPWindow = function (pcpWindow) {
-    var windowSize = pcpWindow.getSize(),
-        width = windowSize[0],
-        height = windowSize[1];
-    pcpWindow.setPosition(-width, -height);
-};
-
-/**
- * Creates a setting view model to be used in the settings window.
- * @param key {String} The name of the setting. Must be unique as
- * subsequent requests to the GPII API will use this key as identifier.
- * @param settingDescriptor {Object} A descriptor for the given setting
- * containing its title, description and constraints regarding its value.
- * @return {Object} The view model for the setting.
- */
-gpii.app.createSettingModel = function (key, settingDescriptor) {
-    var schema = settingDescriptor.schema;
-
-    return {
-        path: key,
-        value: settingDescriptor.value,
-        solutionName: settingDescriptor.solutionName,
-
-        title: schema.title,
-        description: schema.description,
-        icon: "../icons/gear-cloud-white.png",
-        dynamicity: "none", // "none", "application" or "os"
-        isPersisted: false,
-
-        type: schema.type,
-        min: schema.min,
-        max: schema.max,
-        divisibleBy: schema.divisibleBy,
-        values: schema["enum"]
-    };
-};
-
-/**
- * Extracts data for the user's preference sets (including the active preference
- * set and the applicable settings) from the message received when the user keys in.
- * @param message {Object} The message sent when the user keys is (a JSON
- * object).
- * @return {Object} An object containing all preference sets, the active preference
- * set and the corresponding settings.
- */
-gpii.app.extractPreferencesData = function (message) {
-    var value = message.value || {},
-        preferences = value.preferences || {},
-        contexts = preferences.contexts,
-        settingControls = value.settingControls,
-        sets = [],
-        activeSet = value.activeContextName || null,
-        settings = [];
-
-    if (contexts) {
-        sets = fluid.hashToArray(contexts, "path");
-    }
-
-    if (settingControls) {
-        settings = fluid.values(
-            fluid.transform(settingControls, function (settingDescriptor, settingKey) {
-                return gpii.app.createSettingModel(settingKey, settingDescriptor);
-            })
-        );
-    }
-
-    return {
-        sets: sets,
-        activeSet: activeSet,
-        settings: settings
-    };
-};
-
-/**
- * Returns whether the underlying OS is Windows 10 or not.
- * @return {Boolean} `true` if the underlying OS is Windows 10 or
- * `false` otherwise.
- */
-gpii.app.isWin10OS = function () {
-    var osRelease = os.release(),
-        delimiter = osRelease.indexOf("."),
-        majorVersion = osRelease.slice(0, delimiter);
-    return majorVersion === "10";
-};
-
-/**
- * This function takes care of notifying the PCP window whenever the
- * user changes the accent color of the OS theme. Available only if
- * the application is used on Windows 10.
- * @param pcp {Object} The `gpii.app.pcp` instance
- */
-gpii.app.registerAccentColorListener = function (pcp) {
-    if (gpii.app.isWin10OS()) {
-        // Ideally when the PCP window is created, it should be notified about
-        // the current accent color. Possible events which can be used for this
-        // purpose are "ready-to-show" or "show". However, as the window is drawn
-        // off screen, registering the listeners will happen after the corresponding
-        // event has been fired. That is why the PCP window should be notified every
-        // time it is focused (only the first time is insufficient because showing
-        // the window (even off screen) automatically focuses it).
-        pcp.pcpWindow.on("focus", function () {
-            pcp.notifyPCPWindow("accentColorChanged", systemPreferences.getAccentColor());
-        });
-
-        systemPreferences.on("accent-color-changed", function (event, accentColor) {
-            pcp.notifyPCPWindow("accentColorChanged", accentColor);
-        });
-    }
-};
-
-
-/**
- * Register listeners for messages from the GPII socket connection.
- * @param socket {Object} The connected gpii socket
- * @param gpiiConnector {Object} The `gpii.app.gpiiConnector` instance
- * @param pcp {Object} The `gpii.app.pcp` instance
- */
-gpii.app.registerPCPListener = function (socket, gpiiConnector, pcp) {
-    socket.on("message", function (rawData) {
-        var data = JSON.parse(rawData),
-            operation = data.type,
-            path = data.path,
-            preferences;
-
-        if (operation === "ADD") {
-            if (path.length === 0) {
-                /*
-                 * "Keyed in" data has been received
-                 */
-                preferences = gpii.app.extractPreferencesData(data);
-                gpiiConnector.events.onPreferencesUpdated.fire(preferences);
-                pcp.notifyPCPWindow("keyIn", preferences);
-            } else {
-                /*
-                 * Setting change update has been received
-                 */
-                var settingPath = path[path.length - 2],
-                    settingValue = data.value;
-                pcp.notifyPCPWindow("updateSetting", {
-                    path: settingPath,
-                    value: settingValue
-                });
-            }
-        } else if (operation === "DELETE") {
-            preferences = gpii.app.extractPreferencesData(data);
-            gpiiConnector.events.onPreferencesUpdated.fire(preferences);
-            pcp.notifyPCPWindow("keyOut", preferences);
-        }
-    });
-};
-
-/**
- * Opens a connection to the PCP Channel WebSocket.
- * @param config {Object} The configuration for the WebSocket
- */
-gpii.app.createGPIIConnection = function (config) {
-    return new ws(config.gpiiWSUrl); // eslint-disable-line new-cap
-};
-
-
-/**
- * Initialises the connection between the Electron process and
- * the PCP's `BrowserWindow` instance
- *
- * @param pcp {Object} A `gpii.app.pcp` instance
- * @param gpiiConnector {Object} A `gpii.app.gpiiConnector` instance
- */
-gpii.app.initPCPWindowIPC = function (app, pcp, gpiiConnector) {
-    ipcMain.on("closePCP", function () {
-        pcp.hide();
-    });
-
-    ipcMain.on("keyOut", function () {
-        pcp.hide();
-        app.keyOut();
-    });
-
-    ipcMain.on("updateSetting", function (event, arg) {
-        gpiiConnector.updateSetting(arg);
-    });
-
-    ipcMain.on("updateActivePreferenceSet", function (event, arg) {
-        gpiiConnector.updateActivePrefSet(arg.value);
-    });
-
-    ipcMain.on("contentHeightChanged", function (event, contentHeight) {
-        pcp.resize(contentHeight);
-    });
 };
 
 /**
@@ -719,785 +354,6 @@ fluid.onUncaughtException.addListener(function (err) {
         app[0].handleUncaughtException(err);
     }
 }, "gpii.app", "last");
-
-/**
- * Resizes the PCP window and positions it appropriately based on the new height
- * of its content. Makes sure that the window is no higher than the available
- * height of the work area in the primary display. The window will not be resized
- * if its current height is the same as the new height. This behaviour can be
- * overridden using the `forceResize` parameter.
- * @param pcp {Object} A `gpii.app.pcp` instance.
- * @param contentHeight {Number} The new height of the BrowserWindow's content.
- * @param minHeight {Number} The minimum height which the BrowserWindow must have.
- * @param forceResize {Boolean} Whether to resize the window even if the current
- * height of the `BrowserWindow` is the same as the new one. Useful when screen
- * DPI is changed as a result of the application of a user's preferences.
- */
-gpii.app.pcp.resize = function (pcp, contentHeight, minHeight, forceResize) {
-    var pcpWindow = pcp.pcpWindow,
-        wasShown = pcp.isShown(),
-        screenSize = electron.screen.getPrimaryDisplay().workAreaSize,
-        windowSize = pcpWindow.getSize(),
-        windowWidth = windowSize[0],
-        initialHeight = windowSize[1],
-        windowHeight = Math.min(screenSize.height, Math.max(contentHeight, minHeight));
-
-    if (initialHeight === windowHeight && !forceResize) {
-        return;
-    }
-
-    pcpWindow.setSize(windowWidth, windowHeight);
-
-    if (wasShown) {
-        pcp.show();
-    }
-};
-
-/**
- * Handles logic for the PCP window.
- */
-fluid.defaults("gpii.app.pcp", {
-    gradeNames: "fluid.modelComponent",
-
-    model:  {
-        keyedInUserToken: null
-    },
-
-    /*
-     * Options to be passed to the Electron `BrowserWindow`.
-     * Only not reserved options can be altered. See `reservedAttrs`.
-     */
-    attrs: {
-        width: 500,
-        height: 600,
-        show: true,
-        frame: false,
-        fullscreenable: false,
-        resizable: false,
-        alwaysOnTop: true,
-        skipTaskbar: true,
-        backgroundColor: "transparent"
-    },
-
-    members: {
-        pcpWindow: {
-            expander: {
-                funcName: "gpii.app.pcp.makePCPWindow",
-                args: [
-                    "{that}.options.attrs"
-                ]
-            }
-        }
-    },
-    listeners: {
-        "onCreate.initPCPWindowIPC": {
-            listener: "gpii.app.initPCPWindowIPC",
-            args: ["{app}", "{that}", "{gpiiConnector}"]
-        },
-        "onCreate.registerAccentColorListener": {
-            listener: "gpii.app.registerAccentColorListener",
-            args: ["{that}"]
-        },
-        "onCreate.initPCPWindowListeners": {
-            listener: "gpii.app.initPCPWindowListeners",
-            args: ["{that}"]
-        }
-    },
-    invokers: {
-        show: {
-            funcName: "gpii.app.pcp.showPCPWindow",
-            args: ["{that}.pcpWindow"]
-        },
-        hide: {
-            funcName: "gpii.app.pcp.hidePCPWindow",
-            args: ["{that}.pcpWindow"]
-        },
-        isShown: {
-            funcName: "gpii.app.pcp.isPCPWindowShown",
-            args: ["{that}.pcpWindow"]
-        },
-        notifyPCPWindow: {
-            funcName: "gpii.app.pcp.notifyPCPWindow",
-            args: ["{that}.pcpWindow", "{arguments}.0", "{arguments}.1"]
-        },
-        getWindowPosition: {
-            funcName: "gpii.app.getWindowPosition",
-            args: ["{that}.options.attrs.width", "{that}.options.attrs.height"]
-        },
-        resize: {
-            funcName: "gpii.app.pcp.resize",
-            args: ["{that}", "{arguments}.0", "{that}.options.attrs.height", "{arguments}.1"]
-        }
-    }
-});
-
-/**
- * Component that contains an Electron Tray.
- */
-fluid.defaults("gpii.app.tray", {
-    gradeNames: "fluid.modelComponent",
-    members: {
-        tray: {
-            expander: {
-                funcName: "gpii.app.makeTray",
-                args: ["{that}.options.icons.keyedOut", "{pcp}.show"]
-            }
-        }
-    },
-    icons: {
-        keyedIn: "icons/gpii-color.ico",
-        keyedOut: "icons/gpii.ico"
-    },
-    components: {
-        menu: {
-            type: "gpii.app.menuInApp"
-        }
-    },
-    model: {
-        keyedInUserToken: null,
-        icon: "{that}.options.icons.keyedOut",
-        preferences: "{app}.model.preferences",
-        tooltip: ""
-    },
-    modelRelay: {
-        "icon": {
-            target: "icon",
-            source: "{that}.model.keyedInUserToken",
-            singleTransform: {
-                type: "fluid.transforms.valueMapper",
-                defaultInputPath: "",
-                match: [{
-                    inputValue: null,
-                    outputValue: "{that}.options.icons.keyedOut"
-                }],
-                noMatch: "{that}.options.icons.keyedIn"
-            }
-        },
-        "tooltip": {
-            target: "tooltip",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.getTrayTooltip",
-                args: ["{that}.model.preferences", "{that}.options.labels.defaultTooltip"]
-            }
-        }
-    },
-    modelListeners: {
-        "icon": {
-            funcName: "gpii.app.tray.setTrayIcon",
-            args: ["{that}.tray", "{change}.value"],
-            excludeSource: "init"
-        },
-        "tooltip": {
-            "this": "{that}.tray",
-            method: "setToolTip",
-            args: "{that}.model.tooltip"
-        }
-    },
-    labels: {
-        defaultTooltip: "(No one keyed in)"
-    }
-});
-
-/**
-  * Sets the icon for the Electron Tray which represents the GPII application.
-  * @param tray {Object} An instance of an Electron Tray.
-  * @param icon {String} The simple path to the icon file.
-  */
-gpii.app.tray.setTrayIcon = function (tray, icon) {
-    var iconPath = path.join(__dirname, icon);
-    tray.setImage(iconPath);
-};
-
-/**
-  * Creates the Electron Tray
-  * @param icon {String} Path to the icon that represents the GPII in the task tray.
-  * @param openPCP {Function} A function for showing the PCP window. Should be called
-  * whenever the user left clicks on the tray icon or uses the PCP window shortcut.
-  */
-gpii.app.makeTray = function (icon, openPCP) {
-    var tray = new Tray(path.join(__dirname, icon));
-
-    tray.on("click", function () {
-        openPCP();
-    });
-
-    globalShortcut.register("Super+CmdOrCtrl+Alt+U", function () {
-        openPCP();
-    });
-
-    return tray;
-};
-
-/**
- * Returns the tooltip for the Electron Tray based on the active preference set (if any).
- * @param preferences {Object} An object describing the preference sets (including the
- * active one) for the currently keyed-in user (if any).
- * @param defaultTooltip {String} A default tooltip text which should be used in case
- * there is no active preference set.
- * @return The tooltip label for the Electron Tray.
- */
-gpii.app.getTrayTooltip = function (preferences, defaultTooltip) {
-    var activePreferenceSet = fluid.find_if(preferences.sets,
-        function (preferenceSet) {
-            return preferenceSet.path === preferences.activeSet;
-        }
-    );
-    return activePreferenceSet ? activePreferenceSet.name : defaultTooltip;
-};
-
-
-/**
- * Component that contains an Electron Dialog.
- */
-
-fluid.defaults("gpii.app.dialog", {
-    gradeNames: "fluid.modelComponent",
-
-    attrs: {
-        width: 800,
-        height: 600,
-        show: false,
-        frame: false,
-        transparent: true,
-        alwaysOnTop: true,
-        skipTaskBar: true
-    },
-    model: {
-        showDialog: false,
-        dialogMinDisplayTime: 2000, // minimum time to display dialog to user in ms
-        dialogStartTime: 0, // timestamp recording when the dialog was displayed to know when we can dismiss it again
-        timeout: 0
-    },
-    members: {
-        dialog: {
-            expander: {
-                funcName: "gpii.app.makeWaitDialog",
-                args: [
-                    "{that}.options.attrs",
-                    "@expand:{that}.getWindowPosition()"
-                ]
-            }
-        }
-    },
-    modelListeners: {
-        "showDialog": {
-            funcName: "gpii.app.showHideWaitDialog",
-            args: ["{that}", "{change}.value"]
-        }
-    },
-    listeners: {
-        "onDestroy.clearTimers": "gpii.app.clearTimers({that})"
-    },
-    invokers: {
-        getWindowPosition: {
-            funcName: "gpii.app.getWindowPosition",
-            args: [
-                "{that}.options.attrs.width",
-                "{that}.options.attrs.height"
-            ]
-        }
-    }
-});
-
-gpii.app.clearTimers = function (that) {
-    clearTimeout(that.dismissWaitTimeout);
-    clearInterval(that.displayWaitInterval);
-};
-
-/**
- * Creates a dialog. This is done up front to avoid the delay from creating a new
- * dialog every time a new message should be displayed.
- */
-gpii.app.makeWaitDialog = function (windowOptions, position) {
-    var dialog = new BrowserWindow(windowOptions);
-    dialog.setPosition(position.x, position.y);
-
-    var url = fluid.stringTemplate("file://%gpii-app/src/pcp/html/message.html", fluid.module.terms());
-    dialog.loadURL(url);
-    return dialog;
-};
-
-
-gpii.app.showHideWaitDialog = function (that, showDialog) {
-    showDialog ? gpii.app.displayWaitDialog(that) : gpii.app.dismissWaitDialog(that);
-};
-
-/**
- * Shows the dialog on users screen with the message passed as parameter.
- * Records the time it was shown in `dialogStartTime` which we need when
- * dismissing it (checking whether it's been displayed for the minimum amount of time)
- *
- * @param that {Component} the gpii.app instance
- */
-gpii.app.displayWaitDialog = function (that) {
-    that.dialog.show();
-    // Hack to ensure it stays on top, even as the GPII autoconfiguration starts applications, etc., that might
-    // otherwise want to be on top
-    // see amongst other: https://blogs.msdn.microsoft.com/oldnewthing/20110310-00/?p=11253/
-    // and https://github.com/electron/electron/issues/2097
-    if (that.displayWaitInterval) {
-        clearInterval(that.displayWaitInterval);
-        that.displayWaitInterval = 0;
-    }
-    that.displayWaitInterval = setInterval(function () {
-        if (!that.dialog.isVisible()) {
-            clearInterval(that.displayWaitInterval);
-        };
-        that.dialog.setAlwaysOnTop(true);
-    }, 100);
-
-    that.model.dialogStartTime = Date.now();
-};
-
-/**
- * Dismisses the dialog. If less than `that.dialogMinDisplayTime` ms have passed since we first displayed
- * the window, the function waits until `dialogMinDisplayTime` has passed before dismissing it.
- *
- * @param that {Component} the gpii.app instance
- */
-gpii.app.dismissWaitDialog = function (that) {
-    if (that.dismissWaitTimeout) {
-        clearTimeout(that.dismissWaitTimeout);
-        that.dismissWaitTimeout = null;
-    }
-
-    // ensure we have displayed for a minimum amount of `dialogMinDisplayTime` secs to avoid confusing flickering
-    var remainingDisplayTime = (that.model.dialogStartTime + that.model.dialogMinDisplayTime) - Date.now();
-
-    if (remainingDisplayTime > 0) {
-        that.dismissWaitTimeout = setTimeout(function () {
-            that.dialog.hide();
-        }, remainingDisplayTime);
-    } else {
-        that.dialog.hide();
-    }
-};
-
-
-/*
- ** Configuration for using the menu in the app.
- ** Note that this is an incomplete grade which references the app.
- */
-fluid.defaults("gpii.app.menuInApp", {
-    gradeNames: "gpii.app.menu",
-    model: {
-        keyedInUserToken: "{app}.model.keyedInUserToken",
-        preferences: {
-            sets: "{app}.model.preferences.sets",
-            activeSet: "{app}.model.preferences.activeSet"
-        }
-    },
-    modelListeners: {
-        "menuTemplate": {
-            namespace: "updateMenu",
-            funcName: "gpii.app.updateMenu",
-            args: ["{tray}.tray", "{that}.model.menuTemplate", "{that}.events"]
-        }
-    },
-    listeners: {
-        "onPCP.performPCP": {
-            listener: "{pcp}.show"
-        },
-
-        "onActivePrefSetUpdate.performActiveSetChange": {
-            listener: "{gpiiConnector}.updateActivePrefSet",
-            args: "{arguments}.0.path"
-        },
-
-        // onKeyIn event is fired when a new user keys in through the task tray.
-        // This should result in:
-        // 1. key out the old keyed in user token
-        // 2. key in the new user token
-        //   a) trigger GPII {lifecycleManager}.events.onSessionStart
-        //   b) fire a model change to set the new model.keyedInUserToken
-        //   c) update the menu
-        "onKeyIn.performKeyOut": {
-            listener: "{app}.keyOut"
-        },
-        "onKeyIn.performKeyIn": {
-            listener: "{app}.keyIn",
-            args: ["{arguments}.0.token"],
-            priority: "after:performKeyOut"
-        },
-        // onKeyOut event is fired when a keyed-in user keys out through the task tray.
-        // This should result in:
-        // 1. key out the currently keyed in user
-        //    a) change model.keyedInUserToken
-        //    b) update the menu
-        "onKeyOut.performKeyOut": {
-            listener: "{app}.keyOut"
-        },
-
-        // onExit
-        "onExit.performExit": {
-            listener: "{app}.exit"
-        }
-    }
-});
-
-/**
-  * Refreshes the task tray menu for the GPII Application using the menu in the model
-  * @param tray {Object} An Electron 'Tray' object.
-  * @param menuTemplate {Array} A nested array that is the menu template for the GPII Application.
-  * @param events {Object} An object containing the events that may be fired by items in the menu.
-  */
-gpii.app.updateMenu = function (tray, menuTemplate, events) {
-    menuTemplate = gpii.app.menu.expandMenuTemplate(menuTemplate, events);
-
-    tray.setContextMenu(Menu.buildFromTemplate(menuTemplate));
-};
-
-fluid.defaults("gpii.app.menuInAppDev", {
-    gradeNames: "gpii.app.menuInApp",
-    // The list of the default snapsets shown on the task tray menu for key-in
-    snapsets: {
-        label: "{that}.options.menuLabels.keyIn",
-        submenu: [{
-            label: "Larger 125%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_1a"
-            }
-        }, {
-            label: "Larger 150%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_1b"
-            }
-        }, {
-            label: "Larger 175%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_1c"
-            }
-        }, {
-            label: "Dark & Larger 125%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_2a"
-            }
-        }, {
-            label: "Dark & Larger 150%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_2b"
-            }
-        }, {
-            label: "Dark & Larger 175%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_2c"
-            }
-        }, {
-            label: "Read To Me",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_3"
-            }
-        }, {
-            label: "Magnifier 200%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_4a"
-            }
-        }, {
-            label: "Magnifier 400%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_4b"
-            }
-        }, {
-            label: "Magnifier 200% & Display Scaling 175%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_4c"
-            }
-        }, {
-            label: "Dark Magnifier 200%",
-            click: "onKeyIn",
-            args: {
-                token: "snapset_4d"
-            }
-        }, {
-            label: "Multi pref sets; Magnifier",
-            click: "onKeyIn",
-            args: {
-                token: "context1"
-            }
-        }, {
-            label: "Invalid user",
-            click: "onKeyIn",
-            args: {
-                token: "danailbd"
-            }
-        }]
-    },
-    exit: {
-        label: "{that}.options.menuLabels.exit",
-        click: "onExit"
-    }
-});
-
-
-/*
- ** Component to generate the menu tree structure that is relayed to gpii.app for display.
- */
-fluid.defaults("gpii.app.menu", {
-    gradeNames: "fluid.modelComponent",
-    model: {
-        // Expected as configuration
-        //keyedInUserToken: null,
-        //preferences: {
-        //    sets: null,
-        //    activeSet: null
-        //},
-        // locally updated
-        preferenceSetsMenuItems: [],  // Updated on `preferences` changed.
-        keyedInUser: null,            // Must be updated when keyedInUserToken changes.
-        keyOut: null,                 // May or may not be in the menu, must be updated when keyedInUserToken changes.
-        menuTemplate: []              // This is updated on change of keyedInUserToken.
-    },
-    modelRelay: {
-        "userName": {
-            target: "userName",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.menu.getUserName",
-                args: ["{that}.model.keyedInUserToken"]
-            }
-        },
-        "keyedInUser": {
-            target: "keyedInUser",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.menu.getKeyedInUser",
-                args: ["{that}.model.keyedInUserToken", "{that}.model.userName", "{that}.options.menuLabels.keyedIn"]
-            },
-            priority: "after:userName"
-        },
-        "keyOut": {
-            target: "keyOut",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.menu.getKeyOut",
-                args: ["{that}.model.keyedInUserToken", "{that}.options.menuLabels.keyOut", "{that}.options.menuLabels.notKeyedIn"]
-            },
-            priority: "after:userName"
-        },
-        "showPCP": {
-            target: "showPCP",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.menu.getShowPCP",
-                args: ["{that}.model.keyedInUserToken", "{that}.options.menuLabels.pcp"]
-            }
-        },
-        "preferenceSetsMenuItems": {
-            target: "preferenceSetsMenuItems",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.menu.getPreferenceSetsMenuItems",
-                args: ["{that}.model.preferences.sets", "{that}.model.preferences.activeSet"]
-            }
-        },
-        "menuTemplate": {
-            target: "menuTemplate",
-            singleTransform: {
-                type: "fluid.transforms.free",
-                func: "gpii.app.menu.generateMenuTemplate",
-                args: ["{that}.model.showPCP", "{that}.model.keyedInUser", "{that}.options.snapsets", "{that}.model.preferenceSetsMenuItems", "{that}.model.keyOut", "{that}.options.exit"]
-            },
-            priority: "last"
-        }
-    },
-    menuLabels: {
-        pcp: "Open PSP",
-        keyedIn: "Keyed in with %userTokenName",    // string template
-        keyOut: "Key-out of GPII",
-        notKeyedIn: "(No one keyed in)",
-
-        // XXX DEV items
-        exit: "Exit GPII",
-        keyIn: "Key in ..."
-    },
-    events: {
-        onPCP: null,
-        onActivePrefSetUpdate: null,
-        onKeyOut: null,
-
-        // XXX DEV items
-        onKeyIn: null,
-        onExit: null
-    }
-});
-
-/**
-  * Generates a user name to be displayed based on the user token.
-  * @param userToken {String} A user token.
-  */
-gpii.app.menu.getUserName = function (userToken) {
-    // TODO: Name should actually be stored by the GPII along with the user token.
-    var name = userToken ? userToken.charAt(0).toUpperCase() + userToken.substr(1) : "";
-    return name;
-};
-
-
-/**
-*  Object representing options for a `Electron` `ContextMenu` item.
- * @typedef {Object} ElectronMenuItem
- * @property {String} label The label that will be visualized in the menu
- * @property {String} enabled Whether the menu item is enabled
- * @property {String} [type] The type of the menu item
- * @property {String} [click] The event that is fired when the menu item is clicked
- * @property {Object} [args] The arguments to be passed to the click handler
- *               Currently in use
- * @property {String} [args.token] The user token
- * @property {String} [args.path] The path of the setting
- */
-
-/**
-  * Generates an object that represents the menu item for keying in.
-  * @param keyedInUserToken {String} The user token that is currently keyed in.
-  * @param name {String} The name of the user who is keyed in.
-  * @param keyedInStrTemp {String} The string template for the label when a user is keyed in.
-  * @return {ElectronMenuItem}
-  */
-gpii.app.menu.getKeyedInUser = function (keyedInUserToken, name, keyedInStrTemp) {
-    var keyedInUser = null;
-
-    if (keyedInUserToken) {
-        keyedInUser = {
-            label: fluid.stringTemplate(keyedInStrTemp, {"userTokenName": name}),
-            enabled: false
-        };
-    }
-
-    return keyedInUser;
-};
-
-/**
-  * Generates an object that represents the menu item for keying out.
-  * @param keyedInUserToken {String} The user token that is currently keyed in.
-  * @param keyOutStr {String} The string to be displayed for the key out menu item
-  * if there is a keyed in user.
-  * @param notKeyedInStr {String} The string to be displayed when a user is not
-  * keyed in.
-  * @return {ElectronMenuItem}
-  */
-gpii.app.menu.getKeyOut = function (keyedInUserToken, keyOutStr, notKeyedInStr) {
-    var keyOut;
-
-    if (keyedInUserToken) {
-        keyOut = {
-            label: keyOutStr,
-            click: "onKeyOut",
-            args: {
-                token: keyedInUserToken
-            },
-            enabled: true
-        };
-    } else {
-        keyOut = {
-            label: notKeyedInStr,
-            enabled: false
-        };
-    }
-
-    return keyOut;
-};
-
-/**
- * Generates an array that represents the menu items related to a
- * user's preference sets. The returned array can be used in the
- * context menu of a {Tray} object.
- * @param preferenceSets {Array} An array of all preference sets for the user.
- * @param activeSet {String} The path of the currently active preference set.
- * @return {ElectronMenuItem[]}
- */
-gpii.app.menu.getPreferenceSetsMenuItems = function (preferenceSets, activeSet) {
-    var preferenceSetsLabels,
-        separator = {type: "separator"};
-
-    preferenceSetsLabels = preferenceSets.map(function (preferenceSet) {
-        return {
-            label: preferenceSet.name,
-            type: "radio",
-            args: {
-                path: preferenceSet.path
-            },
-            click: "onActivePrefSetUpdate",
-            checked: preferenceSet.path === activeSet
-        };
-    });
-
-    if (preferenceSetsLabels.length > 0) {
-        preferenceSetsLabels.unshift(separator);
-        preferenceSetsLabels.push(separator);
-    }
-
-    return preferenceSetsLabels;
-};
-
-/**
-  * Generates an object that represents the menu items for opening the settings panel
-  * @param keyedInUserToken {String} The user token that is currently keyed in.
-  * @param openSettingsStr {String} The string to be displayed for the open setting panel menu item.
-  * @returns {ElectronMenuItem}
-  */
-gpii.app.menu.getShowPCP = function (keyedInUserToken, openSettingsStr) {
-    return {
-        label: openSettingsStr,
-        click: "onPCP",
-        args: {
-            token: keyedInUserToken
-        }
-    };
-};
-
-/**
-  * Creates a JSON representation of a menu.
-  * @param {Object} An object containing a menu item template.
-  * There should be one object per menu item in the order they should appear in the mneu.
-  */
-gpii.app.menu.generateMenuTemplate = function (/* all the items in the menu */) {
-    var menuTemplate = [];
-    fluid.each(arguments, function (item) {
-        if (item) {
-            if (Array.isArray(item)) {
-                menuTemplate = menuTemplate.concat(item);
-            } else {
-                menuTemplate.push(item);
-            }
-        }
-    });
-
-    return menuTemplate;
-};
-
-/**
-  * Takes a JSON array that represents a menu template and expands the "click" entries into functions
-  * that fire the appropriate event.
-  * @param events {Object} An object that contains the events that might be fired from an item in the menu.
-  * @param menuTemplate {Array} A JSON array that represents a menu template
-  * @return {Array} The expanded menu template. This can be used to create an Electron menu.
-  */
-gpii.app.menu.expandMenuTemplate = function (menuTemplate, events) {
-    fluid.each(menuTemplate, function (menuItem) {
-        if (typeof menuItem.click === "string") {
-            var evtName = menuItem.click;
-            menuItem.click = function () {
-                events[evtName].fire(menuItem.args);
-            };
-        }
-        if (menuItem.submenu) {
-            menuItem.submenu = gpii.app.menu.expandMenuTemplate(menuItem.submenu, events);
-        }
-    });
-
-    return menuTemplate;
-};
-
 
 
 // A wrapper that wraps gpii.app as a subcomponent. This is the grade need by configs/app.json
