+++ src/pcp/index.html
@@ -16,6 +16,11 @@
         <script type="text/javascript" src="js/widgets.js"></script>
         <script type="text/javascript" src="js/exemplars.js"></script>
         <script type="text/javascript" src="js/settingsPanel.js"></script>
+        <script type="text/javascript" src="js/splash.js"></script>
+        <script type="text/javascript" src="js/header.js"></script>
+        <script type="text/javascript" src="js/footer.js"></script>
+        <script type="text/javascript" src="js/clientChannel.js"></script>
+        <script type="text/javascript" src="js/mainWindow.js"></script>
         <script type="text/javascript" src="js/index.js"></script>
     </head>
     <body id="flc-body" class="theme-alt">
