+++ src/pcp/js/settingsPanel.js
@@ -1,21 +1,12 @@
-/*
+/*!
+Copyright 2017 Raising the Floor - International
 
-Copyright 2013-2017 OCAD University
-
-
-
-Licensed under the Educational Community License (ECL), Version 2.0 or the New
-
-BSD license. You may not use this file except in compliance with one these
-
-Licenses.
-
-
-
-You may obtain a copy of the ECL 2.0 License and BSD License at
-
-https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
-
+Licensed under the New BSD license. You may not use this file except in
+compliance with this License.
+The research leading to these results has received funding from the European Union's
+Seventh Framework Programme (FP7/2007-2013) under grant agreement no. 289016.
+You may obtain a copy of the License at
+https://github.com/GPII/universal/blob/master/LICENSE.txt
 */
 
 /* global fluid */
@@ -29,8 +20,8 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
 
     /**
      * Utility function for retrieving the last sub-element of a container
-     * @param container {Object} The jQuery container object
-     * @returns {Object} A jQuery container object
+     * @param container {jQuery} The jQuery container object
+     * @returns {jQuery} A jQuery container object
      */
     gpii.pcp.utils.getContainerLastChild = function (container) {
         return container.children().last();
@@ -38,7 +29,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
 
     /**
      * A simple wrapper for the remove function
-     * @param container {Object} A jQuery object
+     * @param container {jQuery} A jQuery object
      */
     gpii.pcp.utils.removeContainer = function (container) {
         if (container) {
@@ -69,9 +60,8 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
             path: null,
             icon: null,
             solutionName: null,
-            title: null,
-            values: null,
-            value: null
+            value: null,
+            schema: null
         },
         widgetConfig: {
             widgetOptions: null,
@@ -99,7 +89,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
             }, {
                 funcName: "{that}.events.onSettingAltered.fire",
                 args: ["{that}.model.path", "{change}.value"],
-                excludeSource: ["init", "outer"]
+                excludeSource: ["init", "pcp.mainWindow"]
             }]
         },
         listeners: {
@@ -116,7 +106,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
             "onCreate.setTitle": {
                 this: "{that}.dom.title",
                 method: "text",
-                args: "{that}.model.title"
+                args: "{that}.model.schema.title"
             },
             "onCreate.setLabelId": {
                 this: "{that}.dom.titleLabel",
@@ -142,7 +132,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
      */
     gpii.pcp.settingPresenter.updateModelIfNeeded = function (that, path, newValue) {
         if (path === that.model.path) {
-            that.applier.change("value", newValue, null, "outer");
+            that.applier.change("value", newValue, null, "pcp.mainWindow");
         }
     };
 
@@ -151,8 +141,9 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
      * changes a setting which is not dynamic.
      * @param dynamicity {String} The type of dynamicity for the current setting. Can
      * be "none", "application" or "os".
-     * @param restartIcon {Object} A jQuery object representing the restart icon.
-     * @param styles {Object}
+     * @param restartIcon {jQuery} A jQuery object representing the restart icon.
+     * @param styles {Object} An object containing classes to be applied to the restart
+     * icon depending on the setting's dynamicity.
      */
     gpii.pcp.settingPresenter.showRestartIcon = function (dynamicity, restartIcon, styles) {
         if (dynamicity === "application" || dynamicity === "os") {
@@ -167,7 +158,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
      * A method responsible for showing a memory icon if the setting will be persisted
      * after a user has changed it.
      * @param isPersisted {Boolean} Whether the current setting will be persisted or not.
-     * @param memoryIcon {Object} A jQuery object representing the memory icon.
+     * @param memoryIcon {jQuery} A jQuery object representing the memory icon.
      */
     gpii.pcp.settingPresenter.showMemoryIcon = function (isPersisted, memoryIcon) {
         if (isPersisted) {
@@ -257,7 +248,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
                             funcName: "{settingRenderer}.events.onSettingMarkupRendered.fire",
                             /*
                              * Get the widget container.
-                             * Should match single element (jquery returns an array of matches)
+                             * Should match single element (jQuery returns an array of matches)
                              */
                             args: "@expand:$({that}.options.widgetContainerClass, {that}.container)",
                             priority: "after:render"
@@ -401,7 +392,7 @@ https://github.com/fluid-project/infusion/raw/master/Infusion-LICENSE.txt
                         onSettingUpdated: "{settingsPanel}.events.onSettingUpdated"
                     },
 
-                    widgetConfig: "@expand:{settingsVisualizer}.options.widgetExemplars.getExemplarBySchemaType({that}.options.setting.type)",
+                    widgetConfig: "@expand:{settingsVisualizer}.options.widgetExemplars.getExemplarBySchemaType({that}.options.setting.schema.type)",
                     markup: {
                         container: "@expand:gpii.pcp.settingsVisualizer.getIndexedContainerMarkup({settingsVisualizer}.options.dynamicContainerMarkup, {that}.options.settingIndex)",
                         setting: "{settingsVisualizer}.options.markup.setting", // markup.setting",
